refactor(http): move portainer client to client package
